% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpMET-package.R
\docType{package}
\name{harpMET-package}
\alias{harpMET}
\alias{harpMET-package}
\title{harpMET: Interface to harp for MET Norway}
\description{
Includes wrappers to harp functions with some things hard coded for MET Norway
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrew-MET/harpMET}
  \item Report bugs at \url{https://github.com/andrew-MET/harpMET/issues}
}

}
\author{
\strong{Maintainer}: Andrew Singleton \email{andrewts@met.no}

}
\keyword{internal}
