% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_met_analysis.R
\name{read_met_analysis}
\alias{read_met_analysis}
\title{Read 2d fields from the MET Norway 1km Nordic Analysis}
\usage{
read_met_analysis(
  dttm,
  parameter,
  transformation = c("none", "interpolate", "regrid", "xsection", "subgrid"),
  transformation_opts = NULL,
  return_data = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{dttm}{A vector of date time strings to read. Can be in YYYYMMDD,
YYYYMMDDhh, YYYYMMDDhhmm, or YYYYMMDDhhmmss format. Can be numeric or
character.}

\item{parameter}{The name of the forecast parameter(s) to read from the
files. Should either be harp parameter names (see show_harp_parameters), or
in the case of netcdf files can be the name of the parameters in the files.}

\item{transformation}{The transformation to apply to 2d fields. Can be
"interpolate" to interpolate to long-lat locations, "regrid" to regrid the
2d field to a new grid and/or projection, "xsection" to extract a vertical
cross section or "subgrid" to take a subset of the input grid.}

\item{transformation_opts}{Options for the different transformations. See
\link[harpIO]{interpolate_opts} for functions to generate options
appropriate to the selected transformation.}

\item{return_data}{Whether to return the data to session. Set to TRUE by
default unlike \link[harpIO]{read_analysis}. Be careful setting this to
TRUE as it can result in running out of memory.}

\item{show_progress}{Whether to show progress reading the data.}
}
\value{
a harp_analysis object for met_analysis
}
\description{
Calls read_analysis from harpIO with paths for the MET analysis.
}
\examples{
library(harpIO)
read_met_analysis(
  seq_dates(2022081612, 2022081615, "1h"),
  "Pcp"
)

read_met_analysis(
  seq_dates(2022081612, 2022081615, "1h"),
  "T2m"
)
}
